requirements.txt
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.5
bleach==6.2.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.1
cloudscraper==1.2.71
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
dotenv==0.9.9
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
fqdn==1.5.1
fsspec==2025.10.0
google==3.0.0
google-auth==2.43.0
google-genai==1.50.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.10
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
lark==1.2.2
lxml==6.0.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.4
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pandas-stubs==2.3.2.250926
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pgvector==0.4.1
pillow==12.0.0
platformdirs==4.4.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
psutil==7.1.0
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
RateMyProfessor-Database-APIs==1.1.1
RateMyProfessorAPI==1.3.6
RateMyProfessorPyAPI==1.2.1
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
Send2Trash==1.8.3
sentence-transformers==5.1.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
stack-data==0.6.3
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.1
torch==2.9.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
types-python-dateutil==2.9.0.20250822
types-pytz==2025.2.0.20251108
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
